requirements.txt
crewai>=0.11.0
google-cloud-aiplatform>=1.36.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
requests>=2.31.0
beautifulsoup4>=4.12.2
pandas>=2.1.1
numpy>=1.24.3
tavily-python>=0.1.8
google-generativeai>=0.3.0 